fix bug Jan 12 2023

+++ src/lib/AssembleGroup.m
@@ -96,7 +96,7 @@ classdef AssembleGroup < handle
             % Extension length (in the extension direction, opposite from 
             % the module carrying direction): 
             % [left, lower; right, upper];
-            ext_l = gp_loc - [loc; loc];
+            ext_l = gp_loc - [int32(loc); int32(loc)];
             ext_l(1, :) = max(ext_l(1, :), [0, 0]);
             ext_l(2, :) = abs(min(ext_l(2, :), [0, 0]));
             % Combine the cogn map of robot and modules
